+++ llm_bench/python/task/image_generation.py
