+++ llm_bench/python/task/super_resolution_generation.py
